fixup: organize tracing code
